requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langgraph==0.2.28
faiss-cpu==1.8.0
pgvector==0.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.9.2